+++ src/App.tsx
@@ -3,8 +3,8 @@ import './App.css';
 import Home from './components/Home';
 import PlayerBoard from './components/PlayerBoard';
 import ComputerBoard from './components/ComputerBoard';
-import { ComputerBoardContextProvider } from './context/computer_board_context';
-import { PlayerBoardContextProvider } from './context/player_board_context';
+import { ComputerBoardContextProvider } from './context/ComputerBoardContext';
+import { PlayerBoardContextProvider } from './context/PlayerBoardContext';
 
 const App = function () {
   return (
